requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==3.2.15
djangorestframework==3.14.0
idna==3.4
PyJWT==2.6.0
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
stripe==5.0.0
typing-extensions==4.3.0
urllib3==1.26.12
